remove deprecated connector
