+++ sdk/src/EvmWalletConnector.ts
@@ -1,21 +1,4 @@
-import type { AbiMap, Asset, FuelProviderConfig } from '@fuel-wallet/types';
-import {
-  FuelWalletConnector,
-  FuelWalletLocked,
-  FuelWalletProvider,
-  Network
-} from '@fuel-wallet/sdk';
-import {
-  JsonAbi,
-  TransactionRequestLike,
-  Predicate,
-  Address,
-  transactionRequestify,
-  hashTransaction,
-  Provider,
-  InputValue,
-  ProviderOptions
-} from 'fuels';
+// External libraries
 import {
   BytesLike,
   hexlify,
@@ -25,31 +8,126 @@ import {
 import { hexToBytes } from '@ethereumjs/util';
 import memoize from 'memoizee';
 
+import { Asset, AbiMap } from '@fuel-wallet/types';
+import {
+  Provider,
+  transactionRequestify,
+  type TransactionRequestLike,
+  JsonAbi,
+  Predicate,
+  Address,
+  hashTransaction,
+  InputValue
+} from 'fuels';
+import {
+  FuelConnector,
+  Network,
+  Version,
+  ConnectorMetadata
+} from '@fuel-wallet/sdk';
+
 import { EIP1193Provider } from './eip-1193';
 import { getPredicateRoot } from './getPredicateRoot';
 import { predicates } from './predicateResources';
 
-export class EVMWalletConnector extends FuelWalletConnector {
+export class EVMWalletConnector extends FuelConnector {
+  // @ts-expect-error
   ethProvider: EIP1193Provider;
-
-  // Our signer used to interact with the network
+  // @ts-expect-error
   fuelProvider: Provider;
-
   private predicate: { abi: any; bytecode: Uint8Array };
+  private setupLock: boolean = false;
+  private _currentAccount: string | null = null;
 
-  constructor(
-    ethProvider: EIP1193Provider,
-    fuelProvider: Provider,
-    {
-      predicate = 'verification-predicate'
-    }: { predicate?: keyof typeof predicates } = {}
-  ) {
+  // metadata placeholder
+  metadata: ConnectorMetadata = {
+    image: '',
+    install: {
+      action: 'Install',
+      description: 'An EVM wallet is required.',
+      link: ''
+    }
+  };
+
+  constructor() {
     super();
-    this.ethProvider = ethProvider;
-    this.fuelProvider = fuelProvider;
+    this.name = 'EVM wallet connector';
+    this.predicate = predicates['verification-predicate'];
+    this.installed = true;
+  }
 
-    if (!predicates[predicate]) throw new Error('Invalid predicate');
-    this.predicate = predicates[predicate];
+  /**
+   * ============================================================
+   * Application communication methods
+   * ============================================================
+   */
+
+  async getProviders() {
+    if (!this.fuelProvider || !this.ethProvider) {
+      if (typeof window !== 'undefined') {
+        // @ts-expect-error
+        this.ethProvider = window.ethereum;
+        if (!this.ethProvider) {
+          throw new Error('Ethereum provider not found');
+        }
+
+        this.fuelProvider = await Provider.create(
+          'https://beta-4.fuel.network/graphql'
+        );
+        if (!this.fuelProvider) {
+          throw new Error('Fuel provider not found');
+        }
+      } else {
+        throw new Error('window.ethereum not found');
+      }
+    }
+
+    return { fuelProvider: this.fuelProvider, ethProvider: this.ethProvider };
+  }
+
+  async setup() {
+    if (this.setupLock) return;
+    this.setupLock = true;
+    await this.setupCurrentAccount();
+    await this.setupEventBridge();
+  }
+
+  async setupEventBridge() {
+    const { ethProvider } = await this.getProviders();
+    ethProvider.on('accountsChanged', async (accounts) => {
+      this.emit('accounts', await this.accounts());
+      if (this._currentAccount !== accounts[0]) {
+        await this.setupCurrentAccount();
+      }
+    });
+    ethProvider.on('connect', async (arg) => {
+      this.emit('connection', await this.isConnected());
+    });
+    ethProvider.on('disconnect', async (arg) => {
+      this.emit('connection', await this.isConnected());
+    });
+  }
+
+  async setupCurrentAccount() {
+    const [currentAccount = null] = await this.accounts();
+    this._currentAccount = currentAccount;
+    this.emit('currentAccount', currentAccount);
+  }
+
+  /**
+   * ============================================================
+   * Connector methods
+   * ============================================================
+   */
+
+  async ping(): Promise<boolean> {
+    await this.getProviders();
+    await this.setup();
+    return true;
+  }
+
+  async version(): Promise<Version> {
+    return { app: '0.0.0', network: '0.0.0' };
   }
 
   async isConnected(): Promise<boolean> {
@@ -57,23 +135,15 @@ export class EVMWalletConnector extends FuelWalletConnector {
     return accounts.length > 0;
   }
 
-  async connect(): Promise<boolean> {
-    await this.ethProvider.request({ method: 'eth_requestAccounts' });
-    return true;
-  }
-
-  async disconnect(): Promise<boolean> {
-    // TODO: actually disconnect?
-    return true;
-  }
-
-  async accounts(): Promise<string[]> {
+  async accounts(): Promise<Array<string>> {
     // Get the ethereum accounts
-    const ethAccounts: Array<string> = await this.ethProvider.request({
+    const { ethProvider, fuelProvider } = await this.getProviders();
+    const ethAccounts: Array<string> = await ethProvider.request({
       method: 'eth_accounts'
     });
 
-    const chainId = await this.fuelProvider.getChainId();
+    const chainId = fuelProvider.getChainId();
+
     const accounts = ethAccounts.map((account) =>
       getPredicateAddress(
         account,
@@ -86,47 +156,55 @@ export class EVMWalletConnector extends FuelWalletConnector {
     return accounts;
   }
 
-  async currentAccount(): Promise<string> {
+  async connect(): Promise<boolean> {
     if (!(await this.isConnected())) {
-      throw Error('No connected accounts');
+      const { ethProvider } = await this.getProviders();
+      await ethProvider.request({
+        method: 'wallet_requestPermissions',
+        params: [
+          {
+            eth_accounts: {}
+          }
+        ]
+      });
     }
+    this.connected = true;
+    return true;
+  }
 
-    const ethAccounts: string[] = await this.ethProvider.request({
-      method: 'eth_accounts'
-    });
-
-    if (ethAccounts.length === 0) {
-      throw Error('No accounts found');
+  async disconnect(): Promise<boolean> {
+    if (await this.isConnected()) {
+      const { ethProvider } = await this.getProviders();
+      await ethProvider.request({
+        method: 'wallet_revokePermissions',
+        params: [
+          {
+            eth_accounts: {}
+          }
+        ]
+      });
     }
-
-    const fuelAccount = getPredicateAddress(
-      ethAccounts[0]!,
-      await this.fuelProvider.getChainId(),
-      this.predicate.bytecode,
-      this.predicate.abi
-    );
-
-    return fuelAccount;
+    this.connected = false;
+    return true;
   }
 
   async signMessage(address: string, message: string): Promise<string> {
-    // A predicate "account" cannot sign
-    throw new Error('Not implemented');
+    throw new Error('A predicate account cannot sign messages');
   }
 
   async sendTransaction(
     address: string,
-    transaction: TransactionRequestLike,
-    _providerConfig?: FuelProviderConfig,
-    signer?: string
+    transaction: TransactionRequestLike
   ): Promise<string> {
     if (!(await this.isConnected())) {
       throw Error('No connected accounts');
     }
-    const ethAccounts: Array<string> = await this.ethProvider.request({
+
+    const { ethProvider, fuelProvider } = await this.getProviders();
+    const ethAccounts: Array<string> = await ethProvider.request({
       method: 'eth_accounts'
     });
-    const chainId = await this.fuelProvider.getChainId();
+    const chainId = fuelProvider.getChainId();
     const accounts = ethAccounts.map((account) => ({
       ethAccount: account,
       predicateAccount: getPredicateAddress(
@@ -136,24 +214,23 @@ export class EVMWalletConnector extends FuelWalletConnector {
         this.predicate.abi
       )
     }));
-    const account = signer
-      ? accounts.find(({ predicateAccount }) => predicateAccount === signer)
-      : accounts[0];
 
+    const account = accounts.find(
+      ({ predicateAccount }) => predicateAccount === address
+    );
     if (!account) {
-      throw Error('Invalid account');
+      throw Error(`No account found for ${address}`);
     }
-
     const transactionRequest = transactionRequestify(transaction);
 
-    // Create a predicate and set the witness index to call in `main()`
+    // Create a predicate and set the witness index to call in predicate`
     const predicate = createPredicate(
       account.ethAccount,
-      this.fuelProvider,
+      fuelProvider,
       this.predicate.bytecode,
       this.predicate.abi
     );
-    predicate.connect(this.fuelProvider);
+    predicate.connect(fuelProvider);
     predicate.setData(transactionRequest.witnesses.length);
 
     // Attach missing inputs (including estimated predicate gas usage) / outputs to the request
@@ -164,7 +241,7 @@ export class EVMWalletConnector extends FuelWalletConnector {
       predicate.populateTransactionPredicateData(transactionRequest);
 
     const txID = hashTransaction(requestWithPredicateAttached, chainId);
-    const signature = await this.ethProvider.request({
+    const signature = await ethProvider.request({
       method: 'personal_sign',
       params: [txID, account.ethAccount]
     });
@@ -173,14 +250,15 @@ export class EVMWalletConnector extends FuelWalletConnector {
     const compactSignature = splitSignature(hexToBytes(signature)).compact;
 
     // We have a witness, attach it to the transaction for inspection / recovery via the predicate
+    // TODO: is below comment still relevant?
     // TODO: not that there is a strange witness before we add out compact signature
     //       it is [ 0x ] and we may need to update versions later if / when this is fixed
     transactionRequest.witnesses.push(compactSignature);
 
     const transactionWithPredicateEstimated =
-      await this.fuelProvider.estimatePredicates(requestWithPredicateAttached);
+      await fuelProvider.estimatePredicates(requestWithPredicateAttached);
 
-    const response = await this.fuelProvider.operations.submit({
+    const response = await fuelProvider.operations.submit({
       encodedTransaction: hexlify(
         transactionWithPredicateEstimated.toTransactionBytes()
       )
@@ -189,13 +267,30 @@ export class EVMWalletConnector extends FuelWalletConnector {
     return response.submit.id;
   }
 
-  async assets(): Promise<Array<Asset>> {
-    return [];
-  }
+  async currentAccount(): Promise<string | null> {
+    if (!(await this.isConnected())) {
+      throw Error('No connected accounts');
+    }
 
-  async addAsset(asset: Asset): Promise<boolean> {
-    console.warn('A predicate account cannot add an asset');
-    return false;
+    const { ethProvider, fuelProvider } = await this.getProviders();
+    const ethAccounts: string[] = await ethProvider.request({
+      method: 'eth_accounts'
+    });
+
+    if (ethAccounts.length === 0) {
+      throw Error('No accounts found');
+    }
+
+    // Eth Wallet (MetaMask at least) return the current select account as the first
+    // item in the accounts list.
+    const fuelAccount = getPredicateAddress(
+      ethAccounts[0]!,
+      fuelProvider.getChainId(),
+      this.predicate.bytecode,
+      this.predicate.abi
+    );
+
+    return fuelAccount;
   }
 
   async addAssets(assets: Asset[]): Promise<boolean> {
@@ -203,30 +298,35 @@ export class EVMWalletConnector extends FuelWalletConnector {
     return false;
   }
 
-  async getWallet(address: string): Promise<FuelWalletLocked> {
-    const userAccounts = await this.accounts();
-
-    // If the address is not a valid predicate account for this wallet then error
-    if (!userAccounts.includes(address)) {
-      throw Error('Invalid account');
-    }
-
-    const provider = await this.getProvider();
-
-    return new FuelWalletLocked(address, this, provider);
+  async addAsset(asset: Asset): Promise<boolean> {
+    console.warn('A predicate account cannot add an asset');
+    return false;
   }
 
-  async getProvider(): Promise<FuelWalletProvider> {
-    const options: ProviderOptions & { walletConnection: FuelWalletConnector } =
-      {
-        walletConnection: new FuelWalletConnector('EVM-Wallet-Connector')
-      };
+  async assets(): Promise<Array<Asset>> {
+    // TODO: can get assets at a predicates address? emit warning/throw error if not?
+    return [];
+  }
 
-    const walletProvider = await FuelWalletProvider.create(
-      this.fuelProvider.url,
-      options
-    );
-    return walletProvider;
+  async addNetwork(networkUrl: string): Promise<boolean> {
+    console.warn('Cannot add a network');
+    return false;
+  }
+
+  async selectNetwork(_network: Network): Promise<boolean> {
+    // TODO: actually allow selecting networks once mainnet is released?
+    console.warn('Cannot select a network');
+    return false;
+  }
+
+  async networks(): Promise<Network[]> {
+    return [await this.currentNetwork()];
+  }
+
+  async currentNetwork(): Promise<Network> {
+    const { fuelProvider } = await this.getProviders();
+    const chainId = fuelProvider.getChainId();
+    return { url: fuelProvider.url, chainId: chainId };
   }
 
   async addAbi(abiMap: AbiMap): Promise<boolean> {
@@ -239,28 +339,9 @@ export class EVMWalletConnector extends FuelWalletConnector {
   }
 
   async hasAbi(contractId: string): Promise<boolean> {
+    console.warn('A predicate account cannot have an ABI');
     return false;
   }
-
-  async network(): Promise<Network> {
-    const chainId = await this.fuelProvider.getChainId();
-    return { url: this.fuelProvider.url, chainId: chainId };
-  }
-
-  async networks(): Promise<Network[]> {
-    return [await this.network()];
-  }
-
-  async addNetwork(networkUrl: string): Promise<boolean> {
-    throw new Error('Not implemented');
-  }
-
-  // on<E extends FuelEvents['type'], D extends FuelEventArg<E>>(
-  //   eventName: E,
-  //   listener: (data: D) => void
-  // ): this {
-  //   return super.on(eventName, listener);
-  // }
 }
 
 export const getPredicateAddress = memoize(
