+++ sdk/src/index.ts
@@ -1,2 +1,1 @@
 export * from './EvmWalletConnector';
-export { EVMWalletConnectorRefactor } from './ConnectorRefactor';
